Add ability to post photo from NASA everyday @ 1PM
NASA has a nice program called the Astronomy Picture of the Day (APOD)
that shares interesting pictures and their description. They also have a
cool API to retrieve the day's picture.

The new code introduces our bot to the NASA APOD API and allows it to
fetch the day's photo and post to the #random channel at 1PM everyday.